Move tests to `Functional`
... as they are co-dependent on `OutputFormat`.

Also, add tests for `createCompact` and `createPretty` flavours of
`OutputFormat`.
